Implementation of JeOS Operator
This commit introduces an operating system configuration controller
for SUSE JeOS. The controller is based in alicloud's coreos
controller. The common code is moved to a generic oscommon controller
which implement the bulk of the logic. JeOS controller implementation
is limited to provide a template for rendering the cloud-init
configuration.

Signed-off-by: Pablo Chacin <pchacin@suse.com>