Make webhook server port configurable and default to 443
By default, GKE only allows connections from its kube-apiservers to pods
on port 443 and 10250. If you want to allow a different port you have to
add a firewall rule for your GKE cluster. To prevent this we set the
default port to 443 now to make the extensions useable out-of-the-box on
GKE clusters.