Scale down/up cluster-autoscaler instead of deleting it
Most parts of the code in the generic worker actuator have been copied from the code in Gardener that was written over a year ago and before we started extensibility.
Back then when Gardener handled the machines it could delete the cluster-autoscaler deployment during a rolling update of worker machines because it knew that the task that would re-deploy the cluster-autoscaler would be executed after the rolling update has completed successfully.
Now, with extensibility, the reconciliation of the worker actuator and Gardener might not be in sync. Consequently, it might happen that the worker actuator reconciles and deletes  the cluster-autoscaler, but Gardener does not reconcile in order to bring it back (only later).
Let's improve the code such that the cluster-autoscaler deployment is not deleted by the worker actuator, instead, it's just scaled down and scaled up again at the end (without the necessity to wait for Gardener to do the same).