Control plane generic actuator waits for managed resources to be deleted
We have seen situations in which the shoot namespace in the seed was
hanging in "Terminating" state because the managed resources from the
control plane controller were still existing.
In the [shoot deletion flow](https://github.com/gardener/gardener/blob/master/pkg/controllermanager/controller/shoot/shoot_control_delete.go)
the `ControlPlane` resource is destroyed, and right after that the
kube-apiserver deployment is deleted. If the gardener-resource-manager
is not cleaning up the control plane MRs fast enough then the
kube-apiserver is deleted to fast which will block the
gardener-resource-manager from gracefully cleaning up the MRs.
Thus, now with the waiting mechanism we ensure that the `ControlPlane`
resource is only deleted after the MRs are gone, hence, the
kube-apiserver deployment will also only be deleted after that.