Azure: Update deprecated terraform resources
The bindings of route tables and security groups directly to subnets is deprecated and will be removed in the next major version of the Azurerm terraform provider.
As replacement there are new association resources to bind route tables and security groups to a subnet. Migration will work out of the box.