Rename calico-typha deployment to enable updating existing shoots
We changed the name of the default deployment name from "calico-typha" to "calico-typha-deploy" because
Gardener deployed calico-typha with 0 replicas. Now, we don't specify `.spec.replicas` anymore (because we
use the typha-HPA for auto-scaling), hence, for existing clusters the replicas will never get `1` (because
the typha-HPA is a new pod that cannot start without typha to run).

Co-Authored-By: zanetworker <adel.zalok.89@gmail.com>